Changes to Tracker
I made changes to two constants that that should be helpful, the first was that the window that determined what an acceptable jump between a frame was, it was set to 500pxs that was probably way too high, gonna use 100px for now, the second was how many frames a cell would look for matches, before it would search for 3 frames, now it searches for 5.

So what was basically happening was whenever it got occluded it would look for a better pair, even one way too far out and would pick it up because it had no better options and because it was allowed to.

I also split up the makin of cost edges and assigning tracks into two different loops to keep things looking nice and get some speed.